remove batch endpoint
