+++ crates/test-utils/src/store/issuer.json
@@ -1,7 +1,6 @@
 {
     "credential_issuer": "http://vercre.io",
     "credential_endpoint": "http://vercre.io/credential",
-    "batch_credential_endpoint": "http://vercre.io/batch",
     "deferred_credential_endpoint": "http://vercre.io/deferred",
     "display": {
         "name": "Vercre",
